--- src/components/Home.js
@@ -1,22 +1,0 @@
-import React from 'react'
-
-function Home() {
-    return (
-        <div id='home-page-wrapper'>
-            <div id='home-section-1' className='mt-4'>
-                <img 
-                    src='../images/stock-images/stock_img1.jpg' 
-                    alt='bedroom'
-                    className='w-100'
-                />
-                <img 
-                    src='../images/stock-images/stock_img1.jpg' 
-                    alt='bedroom'
-                    className='w-100'
-                />
-            </div>
-        </div>
-    )
-}
-
-export default Home
