+++ src/components/Home/Home.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import HomeSection1 from './HomeSection1'
+
+function Home() {
+    return (
+        <div id='home-page-wrapper'>
+            <HomeSection1 />
+        </div>
+    )
+}
+
+export default Home
